// src/types/pdd-project.ts
export interface PDDProject {
  id: string;
  title: string;
  userId: string;
  userName: string;
  userEmail: string;
  userPhotoURL?: string;
  createdAt: any;
  instagramUrl?: string;
  linkedinUrl?: string;
  githubUrl?: string;
  websiteUrl?: string;
  gpcuDocUrl?: string;
  patentDocUrl?: string;
  canvaUrl?: string;
  figmaUrl?: string;
  gslidesUrl?: string;
}